feat: rename oldgame.ts to GamePage.ts and fix LobbyPage for opponent avatar
